requirements.txt
asgiref==3.3.1
boto3==1.17.27
botocore==1.20.27
Django==3.1.7
django-tinymce4-lite==1.8.0
django-tinymce4-widget==6.1.0
jmespath==0.10.0
jsmin==2.2.2
Pillow==8.1.2
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
